ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Kafka
confluent-kafka==2.3.0
aiokafka==0.10.0

# Redis
redis==5.0.0

# ML/Data Science
numpy==1.26.0
pandas==2.1.0
scikit-learn==1.4.0
joblib==1.3.0

# MLflow
mlflow==2.10.0

# HTTP client
httpx==0.26.0
aiohttp==3.9.0

# Utilities
python-json-logger==2.0.7
structlog==24.1.0
prometheus-client==0.19.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pgvector==0.3.5
# Keep AI image lean on local/dev Docker by avoiding CUDA wheel dependencies.
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1+cpu
sentence-transformers==3.0.1

# Gemini
google-genai
